+++ TexturedQuad.cpp
@@ -212,9 +212,9 @@ void TexturedQuad::CreateSurfaceTexture()
     {
         for (int y = 0; y < m_surface.Height(); ++y)
         {
-            float u = static_cast<float>(x) / 10.f;
-            float v = static_cast<float>(y) / 10.f;
-            float n = noise(u, v);
+            double u = static_cast<double>(x) / 10.f;
+            double v = static_cast<double>(y) / 10.f;
+            double n = noise(u, v);
             Uint8 c = static_cast<Uint8>(std::clamp(static_cast<int>(255.f * n), 0, 255));
             Uint32 color = RGB(c, c, c);
             m_surface.SetPixel(x, y, color);
